arrrghhhhhhhhh async != defer
soooo it turns out that Chrome is so efficient that it will async load
and execute the entire application.js file before the page has even
finished rendering. this is less than ideal when you need the code
inside application.js to run, say, onLoad. when you need your code to
run after the DOM has been built, you're supposed to set defer=true, so
that the browser knows it can render as fast as possible and then run
your code after that. which it now does. which means everything
(including the Iconic icons) now works in both Safari and
annoyingly-fast Chrome.