We can avoid being defensive if we just check for an invoice.
The previous code only existed for one-off payments, which was treated
as an edge case in this way.  Now, we just check for an invoice inside
the charge record and if one exists, they are a user with a subscription,
and otherwise it is assumed they are a one-off and we will short circuit.