fix links to membership#destroy, put with "other"
it turns out that singular resources (like `resource :membership`
instead of `resource :memberships`) don't accept a record as an
argument. if you try to pass a record as an argument anyway, rails will
assume that the id of the record you passed is supposed to be the format
of the URL, and generate urls like `/membership.ID_HERE`. removing the
record will generate URLs that point to the singular membership, like
`/membership`, which is what we want in this case.