switch from Membership#kind to #level
Level is independent of monthly/yearly, and counts legacy members as the
level that they are currently at, not the level they started at and are
still paying for.

It's useless for calculating income, but so was #kind. This tries to
make that lack of connection between level and income more obvious.